+++ ibllib/atlas/flatmaps.py
@@ -152,8 +152,8 @@ def swanson_json(filename="swansonpaths.json"):
 
 
 def plot_swanson_vector(acronyms=None, values=None, ax=None, hemisphere=None, br=None, orientation='landscape',
-                        empty_color='silver', vmin=None, vmax=None, cmap='viridis', annotate=False, mask=None, mask_color='w',
-                        **kwargs):
+                        empty_color='silver', vmin=None, vmax=None, cmap='viridis', annotate=False, annotate_n=10,
+                        annotate_order='top', annotate_list=None, mask=None, mask_color='w', fontsize=10, **kwargs):
 
     br = BrainRegions() if br is None else br
     br.compute_hierarchy()
@@ -178,6 +178,8 @@ def plot_swanson_vector(acronyms=None, values=None, ax=None, hemisphere=None, br
     sw = swanson()
     sw_json = swanson_json()
 
+    plot_idx = []
+    plot_val = []
     for i, reg in enumerate(sw_json):
 
         if acronyms is None:
@@ -185,6 +187,8 @@ def plot_swanson_vector(acronyms=None, values=None, ax=None, hemisphere=None, br
         else:
             idx = np.where(ibr == reg['thisID'])[0]
             if len(idx) > 0:
+                plot_idx.append(ibr[idx[0]])
+                plot_val.append(vals[idx[0]])
                 color = rgba_color[idx[0]] / 255
             else:
                 idx = np.where(imr == reg['thisID'])[0]
@@ -247,12 +251,27 @@ def plot_swanson_vector(acronyms=None, values=None, ax=None, hemisphere=None, br
             ax.set_ylim(0, 2 * sw.shape[0])
 
     if annotate:
-        annotate_swanson(ax=ax, orientation=orientation, br=br)
+        if annotate_list is not None:
+            annotate_swanson(ax=ax, acronyms=annotate_list, orientation=orientation, br=br, thres=10, fontsize=fontsize)
+        elif acronyms is not None:
+            ids = br.index2id(np.array(plot_idx))
+            _, indices, _ = np.intersect1d(br.id, br.remap(ids, 'Swanson-lr'), return_indices=True)
+            a, b = ismember(ids, br.id[indices])
+            sorted_id = ids[a]
+            vals = np.array(plot_val)[a]
+            sort_vals = np.argsort(vals) if annotate_order == 'bottom' else np.argsort(vals)[::-1]
+            annotate_swanson(ax=ax, acronyms=sorted_id[sort_vals[:annotate_n]], orientation=orientation, br=br,
+                             thres=10, fontsize=fontsize)
+        else:
+            annotate_swanson(ax=ax, orientation=orientation, br=br, fontsize=fontsize)
 
     def format_coord(x, y):
-        ind = sw[int(y), int(x)]
-        ancestors = br.ancestors(br.id[ind])['acronym']
-        return f'sw-{ind}, {ancestors}, aid={br.id[ind]}-{br.acronym[ind]} \n {br.name[ind]}'
+        try:
+            ind = sw[int(y), int(x)]
+            ancestors = br.ancestors(br.id[ind])['acronym']
+            return f'sw-{ind}, {ancestors}, aid={br.id[ind]}-{br.acronym[ind]} \n {br.name[ind]}'
+        except IndexError:
+            return ''
 
     ax.format_coord = format_coord
 
@@ -331,12 +350,12 @@ def plot_swanson(acronyms=None, values=None, ax=None, hemisphere=None, br=None,
 
 
 @lru_cache(maxsize=None)
-def _swanson_labels_positions():
+def _swanson_labels_positions(thres=20000):
     """
     This functions computes label positions to overlay on the Swanson flatmap
     :return: dictionary where keys are acronyms
     """
-    NPIX_THRESH = 20000  # number of pixels above which region is labeled
+    NPIX_THRESH = thres  # number of pixels above which region is labeled
     s2a = swanson()
     iw, ih = np.meshgrid(np.arange(s2a.shape[1]), np.arange(s2a.shape[0]))
     # compute the center of mass of all regions (fast enough to do on the fly)
@@ -369,7 +388,7 @@ def _swanson_labels_positions():
     return labels
 
 
-def annotate_swanson(ax, acronyms=None, orientation='landscape', br=None, **kwargs):
+def annotate_swanson(ax, acronyms=None, orientation='landscape', br=None, thres=20000, **kwargs):
     """
     Display annotations on the flatmap
     :param ax:
@@ -385,11 +404,11 @@ def annotate_swanson(ax, acronyms=None, orientation='landscape', br=None, **kwar
     else:  # tech debt: here in fact we should remap and compute labels for hierarchical regions
         aids = br.parse_acronyms_argument(acronyms)
         _, indices, _ = np.intersect1d(br.id, br.remap(aids, 'Swanson-lr'), return_indices=True)
-    labels = _swanson_labels_positions()
+    labels = _swanson_labels_positions(thres=thres)
     for ilabel in labels:
         # do not display uwanted labels
         if ilabel not in indices:
             continue
         # rotate the labels if the dislay is in portrait mode
-        xy = reversed(labels[ilabel]) if orientation == 'portrait' else labels[ilabel]
+        xy = np.flip(labels[ilabel]) if orientation == 'portrait' else labels[ilabel]
         ax.annotate(br.acronym[ilabel], xy=xy, ha='center', va='center', **kwargs)